+++ feature_tests/kotlin/somelib/src/main/kotlin/dev/diplomattest/somelib/OpaqueMutexedString.kt
@@ -18,10 +18,10 @@ internal interface OpaqueMutexedStringLib: Library {
 
 class OpaqueMutexedString internal constructor (
     internal val handle: Pointer,
-
     // These ensure that anything that is borrowed is kept alive and not cleaned
     // up by the garbage collector.
-    internal val selfEdges: List<Any>) {
+    internal val selfEdges: List<Any>
+)  {
 
     internal class OpaqueMutexedStringCleaner(val handle: Pointer, val lib: OpaqueMutexedStringLib) : Runnable {
         override fun run() {
@@ -33,9 +33,9 @@ class OpaqueMutexedString internal constructor (
         internal val libClass: Class<OpaqueMutexedStringLib> = OpaqueMutexedStringLib::class.java
         internal val lib: OpaqueMutexedStringLib = Native.load("somelib", libClass)
         
-        fun fromUsize(number: Long): OpaqueMutexedString {
-            val returnVal = lib.OpaqueMutexedString_from_usize(number);
-        val selfEdges: List<Any> = listOf()
+        fun fromUsize(number: ULong): OpaqueMutexedString {
+            val returnVal = lib.OpaqueMutexedString_from_usize(number.toLong());
+            val selfEdges: List<Any> = listOf()
             val handle = returnVal 
             val returnOpaque = OpaqueMutexedString(handle, selfEdges)
             CLEANER.register(returnOpaque, OpaqueMutexedString.OpaqueMutexedStringCleaner(handle, OpaqueMutexedString.lib));
@@ -45,21 +45,21 @@ class OpaqueMutexedString internal constructor (
         
         fun borrowOther(other: OpaqueMutexedString): OpaqueMutexedString {
             val returnVal = lib.OpaqueMutexedString_borrow_other(other.handle);
-        val selfEdges: List<Any> = listOf(other)
+            val selfEdges: List<Any> = listOf(other)
             val handle = returnVal 
             val returnOpaque = OpaqueMutexedString(handle, selfEdges)
             return returnOpaque
         }
     }
     
-    fun change(number: Long): Unit {
-        val returnVal = lib.OpaqueMutexedString_change(handle, number);
-    
+    fun change(number: ULong): Unit {
+        val returnVal = lib.OpaqueMutexedString_change(handle, number.toLong());
+        
     }
     
     fun borrow(): OpaqueMutexedString {
         val returnVal = lib.OpaqueMutexedString_borrow(handle);
-    val selfEdges: List<Any> = listOf(this)
+        val selfEdges: List<Any> = listOf(this)
         val handle = returnVal 
         val returnOpaque = OpaqueMutexedString(handle, selfEdges)
         return returnOpaque
@@ -67,25 +67,25 @@ class OpaqueMutexedString internal constructor (
     
     fun borrowSelfOrOther(other: OpaqueMutexedString): OpaqueMutexedString {
         val returnVal = lib.OpaqueMutexedString_borrow_self_or_other(handle, other.handle);
-    val selfEdges: List<Any> = listOf(this) + listOf(other)
+        val selfEdges: List<Any> = listOf(this) + listOf(other)
         val handle = returnVal 
         val returnOpaque = OpaqueMutexedString(handle, selfEdges)
         return returnOpaque
     }
     
-    fun getLenAndAdd(other: Long): Long {
-        val returnVal = lib.OpaqueMutexedString_get_len_and_add(handle, other);
-    return returnVal
+    fun getLenAndAdd(other: ULong): ULong {
+        val returnVal = lib.OpaqueMutexedString_get_len_and_add(handle, other.toLong());
+        return returnVal.toULong()
     }
     
     fun dummyStr(): String {
         val returnVal = lib.OpaqueMutexedString_dummy_str(handle);
-        return PrimitiveArrayTools.getUtf8(returnVal)
+            return PrimitiveArrayTools.getUtf8(returnVal)
     }
     
     fun wrapper(): Utf16Wrap {
         val returnVal = lib.OpaqueMutexedString_wrapper(handle);
-    val selfEdges: List<Any> = listOf()
+        val selfEdges: List<Any> = listOf()
         val handle = returnVal 
         val returnOpaque = Utf16Wrap(handle, selfEdges)
         CLEANER.register(returnOpaque, Utf16Wrap.Utf16WrapCleaner(handle, Utf16Wrap.lib));
